feat: share form field logic and ui between apps
