+++ apps/mobile/src/components/FormFields/ActivityTypeSelector/ActivityTypeSelector.stories.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
+import { useForm } from "@routine-support/forms";
 import { storiesOf } from "@storybook/react-native";
-import { useForm } from "react-hook-form";
 
 import { ActivityTypeSelector } from "./ActivityTypeSelector";
