+++ apps/mobile/src/components/FormFields/ClockTypeSelector/ClockTypeSelector.stories.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
+import { useForm } from "@routine-support/forms";
 import { storiesOf } from "@storybook/react-native";
-import { useForm } from "react-hook-form";
 
 import { ClockTypeSelector } from "./ClockTypeSelector";
