+++ apps/mobile/src/components/FormFields/Controller/Controller.tsx
@@ -1,47 +1,5 @@
-import React, { ReactElement } from "react";
-
-import {
-  ControllerFieldState,
-  ControllerRenderProps,
-  Controller as FormController,
-  UseFormStateReturn,
-} from "react-hook-form";
-import { useTranslation } from "react-i18next";
+import { createController } from "@routine-support/forms";
 
 import { FieldWrapper } from "../FieldWrapper";
-import { FormFieldProps } from "../types";
 
-interface ControllerProps extends FormFieldProps {
-  render: (arg: {
-    field: ControllerRenderProps;
-    fieldState: ControllerFieldState;
-    formState: UseFormStateReturn<unknown>;
-  }) => ReactElement;
-  row?: boolean;
-}
-
-export const Controller: React.FC<ControllerProps> = ({
-  name,
-  control,
-  render,
-  label,
-  required,
-}) => {
-  const { t } = useTranslation();
-
-  return (
-    <FormController
-      name={name}
-      control={control}
-      rules={{ required: required && t<string>("Required field") }}
-      render={({ field, fieldState, formState }) => (
-        <FieldWrapper
-          label={label && t<string>(label)}
-          error={fieldState.error?.message && t<string>(fieldState.error?.message)}
-        >
-          {render({ field, fieldState, formState })}
-        </FieldWrapper>
-      )}
-    />
-  );
-};
+export const Controller = createController(FieldWrapper);
