+++ apps/mobile/src/components/FormFields/PictogramSelector/PictogramSelector.stories.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 
+import { useForm } from "@routine-support/forms";
 import { PICTOGRAMS } from "@routine-support/pictograms";
 import { storiesOf } from "@storybook/react-native";
-import { useForm } from "react-hook-form";
 
 import { PictogramSelector } from "./PictogramSelector";
