+++ apps/web/src/components/FormFields/FieldWrapper/index.ts
@@ -1,0 +1,1 @@
+export { FieldWrapper } from "./FieldWrapper";
