+++ apps/web/src/components/FormFields/Select/Select.tsx
@@ -1,11 +1,12 @@
 import React from "react";
 
+import { FormFieldProps } from "@routine-support/forms";
+
 import {
   Select as UncontrolledSelect,
   SelectProps as UncontrolledSelectProps,
 } from "../../Select";
 import { Controller } from "../Controller";
-import { FormFieldProps } from "../types";
 
 type SelectProps = FormFieldProps & UncontrolledSelectProps;
