+++ libs/forms/src/components/Controller/index.ts
@@ -1,0 +1,1 @@
+export { createController } from "./Controller";
