+++ libs/forms/src/types/index.ts
@@ -1,2 +1,4 @@
+export type { FieldWrapperProps } from "./FieldWrapperProps";
 export type { FormError } from "./FormError";
+export type { FormFieldProps } from "./FormFieldProps";
 export type { SubmitErrorData } from "./SubmitErrorData";
